fix: bring observedGeneration and workloadGeneration to match k8s type convention

+++ utils/annotations/annotations.go
@@ -38,20 +38,20 @@ func GetDesiredReplicasAnnotation(rs *appsv1.ReplicaSet) (int32, bool) {
 }
 
 // GetWorkloadGenerationAnnotation returns generation of referenced workload
-func GetWorkloadGenerationAnnotation(ro *v1alpha1.Rollout) (int32, bool) {
+func GetWorkloadGenerationAnnotation(ro *v1alpha1.Rollout) (int64, bool) {
 	if ro == nil {
 		return 0, false
 	}
 	annotationValue, ok := ro.Annotations[WorkloadGenerationAnnotation]
 	if !ok {
-		return int32(0), false
+		return int64(0), false
 	}
 	intValue, err := strconv.ParseInt(annotationValue, 10, 32)
 	if err != nil {
 		log.Warnf("Cannot convert the value %q with annotation key %q for the replica set %q", annotationValue, WorkloadGenerationAnnotation, ro.Name)
-		return int32(0), false
+		return int64(0), false
 	}
-	return int32(intValue), true
+	return intValue, true
 }
 
 // GetRevisionAnnotation returns revision of rollout
